+++ zero.core/patterns/semaphore/IoZeroValMut.cs
@@ -161,7 +161,7 @@ namespace zero.core.patterns.semaphore
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroHead()
+        long IIoZeroSemaphore.ZeroHead()
         {
             throw new NotImplementedException();
         }
@@ -181,22 +181,22 @@ namespace zero.core.patterns.semaphore
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroNextHead()
+        long IIoZeroSemaphore.ZeroNextHead()
         {
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroNextTail()
+        long IIoZeroSemaphore.ZeroNextTail()
         {
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroPrevHead()
+        long IIoZeroSemaphore.ZeroPrevHead()
         {
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroPrevTail()
+        long IIoZeroSemaphore.ZeroPrevTail()
         {
             throw new NotImplementedException();
         }
@@ -206,7 +206,7 @@ namespace zero.core.patterns.semaphore
             throw new NotImplementedException();
         }
 
-        int IIoZeroSemaphore.ZeroTail()
+        long IIoZeroSemaphore.ZeroTail()
         {
             throw new NotImplementedException();
         }
